Creating backend with Ktor
Simple backend which consist in 4 endpoints, one to give a greeting message and the other 3 for manage attendees to an event, the database holds data in app memory so anytime the services dies the data dies with it.

Endpoints handle custom serialization with GSON although it can be do it with Jackson or a custom client.

Simple error handling was also added, there are other ways but for now this version is enough